Make patching API public
