+++ check_api/src/main/java/com/google/errorprone/BaseErrorProneJavaCompiler.java
@@ -19,24 +19,15 @@ package com.google.errorprone;
 import static com.google.common.base.StandardSystemProperty.JAVA_SPECIFICATION_VERSION;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableSet;
-import com.google.errorprone.RefactoringCollection.RefactoringResult;
-import com.google.errorprone.scanner.ErrorProneScannerTransformer;
 import com.google.errorprone.scanner.ScannerSupplier;
 import com.sun.source.util.JavacTask;
-import com.sun.source.util.TaskEvent;
-import com.sun.source.util.TaskEvent.Kind;
-import com.sun.source.util.TaskListener;
 import com.sun.tools.javac.api.BasicJavacTask;
 import com.sun.tools.javac.api.JavacTool;
 import com.sun.tools.javac.util.Context;
 import com.sun.tools.javac.util.JavacMessages;
-import com.sun.tools.javac.util.Log;
-import com.sun.tools.javac.util.Log.WriterKind;
 import com.sun.tools.javac.util.Options;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.io.PrintWriter;
 import java.io.Writer;
 import java.nio.charset.Charset;
 import java.util.EnumSet;
@@ -94,9 +85,9 @@ public class BaseErrorProneJavaCompiler implements JavaCompiler {
     setupMessageBundle(context);
     RefactoringCollection[] refactoringCollection = {null};
     javacTask.addTaskListener(
-        createAnalyzer(scannerSupplier, errorProneOptions, context, refactoringCollection));
+        ErrorProneAnalyzer.createAnalyzer(scannerSupplier, errorProneOptions, context, refactoringCollection));
     if (refactoringCollection[0] != null) {
-      javacTask.addTaskListener(new RefactoringTask(context, refactoringCollection[0]));
+      javacTask.addTaskListener(new ErrorProneAnalyzer.RefactoringTask(context, refactoringCollection[0]));
     }
   }
 
@@ -205,72 +196,5 @@ public class BaseErrorProneJavaCompiler implements JavaCompiler {
   public static void setupMessageBundle(Context context) {
     ResourceBundle bundle = ResourceBundle.getBundle("com.google.errorprone.errors");
     JavacMessages.instance(context).add(l -> bundle);
-  }
-
-  static ErrorProneAnalyzer createAnalyzer(
-      ScannerSupplier scannerSupplier,
-      ErrorProneOptions epOptions,
-      Context context,
-      RefactoringCollection[] refactoringCollection) {
-    if (!epOptions.patchingOptions().doRefactor()) {
-      return ErrorProneAnalyzer.createByScanningForPlugins(scannerSupplier, epOptions, context);
-    }
-    refactoringCollection[0] = RefactoringCollection.refactor(epOptions.patchingOptions(), context);
-
-    // Refaster refactorer or using builtin checks
-    CodeTransformer codeTransformer =
-        epOptions
-            .patchingOptions()
-            .customRefactorer()
-            .or(
-                () -> {
-                  ScannerSupplier toUse = ErrorPronePlugins.loadPlugins(scannerSupplier, context);
-                  ImmutableSet<String> namedCheckers = epOptions.patchingOptions().namedCheckers();
-                  if (!namedCheckers.isEmpty()) {
-                    toUse = toUse.filter(bci -> namedCheckers.contains(bci.canonicalName()));
-                  } else {
-                    toUse = toUse.applyOverrides(epOptions);
-                  }
-                  return ErrorProneScannerTransformer.create(toUse.get());
-                })
-            .get();
-
-    return ErrorProneAnalyzer.createWithCustomDescriptionListener(
-        codeTransformer, epOptions, context, refactoringCollection[0]);
-  }
-
-  static class RefactoringTask implements TaskListener {
-
-    private final Context context;
-    private final RefactoringCollection refactoringCollection;
-
-    public RefactoringTask(Context context, RefactoringCollection refactoringCollection) {
-      this.context = context;
-      this.refactoringCollection = refactoringCollection;
-    }
-
-    @Override
-    public void started(TaskEvent event) {}
-
-    @Override
-    public void finished(TaskEvent event) {
-      if (event.getKind() != Kind.GENERATE) {
-        return;
-      }
-      RefactoringResult refactoringResult;
-      try {
-        refactoringResult = refactoringCollection.applyChanges(event.getSourceFile().toUri());
-      } catch (Exception e) {
-        PrintWriter out = Log.instance(context).getWriter(WriterKind.ERROR);
-        out.println(e.getMessage());
-        out.flush();
-        return;
-      }
-      if (refactoringResult.type() == RefactoringCollection.RefactoringResultType.CHANGED) {
-        PrintWriter out = Log.instance(context).getWriter(WriterKind.NOTICE);
-        out.println(refactoringResult.message());
-        out.flush();
-      }
-    }
   }
 }
