+++ check_api/src/main/java/com/google/errorprone/RefactoringCollection.java
@@ -50,7 +50,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 /** A container of fixes that have been collected during a single compilation phase. */
-class RefactoringCollection implements DescriptionListener.Factory {
+public class RefactoringCollection implements DescriptionListener.Factory {
 
   private static final Logger logger = Logger.getLogger(RefactoringCollection.class.getName());
